+++ frontend/src/layouts/GlobalLayout/GlobalLayout.tsx
@@ -6,23 +6,21 @@ import { AuthProvider } from '@contexts/AuthProvider';
 
 import ErrorPage from '@pages/ErrorPage';
 
-import Header from '@components/_common/Header';
 import PageMoveLoading from '@components/_common/PageMoveLoading';
 
-import { s_content, s_globalContainer } from './GlobalLayout.styles';
+import { s_globalContainer } from './GlobalLayout.styles';
 
 export default function GlobalLayout() {
   return (
     <AuthProvider>
       <div css={s_globalContainer}>
-        <Header />
-        <div css={s_content}>
+        <Suspense fallback={<PageMoveLoading />}>
           <ErrorBoundary fallback={({ error }) => <ErrorPage error={error} />}>
             <Suspense fallback={<PageMoveLoading />}>
               <Outlet />
             </Suspense>
           </ErrorBoundary>
-        </div>
+        </Suspense>
       </div>
     </AuthProvider>
   );
